dev-requirements.txt
-r requirements.txt

factory-boy==2.7.0
fake-factory==0.7.2

pytest-benchmark==3.0.0
pytest==3.0.6
# pulling from github because master has django_assert_num_queries context mgr
git+https://github.com/pytest-dev/pytest-django.git@master
